fix: add END_BLOCK controller call
Also, remove the `action.committed` field; it is a red herring.
